gráfico hora ok, comparação ainda não
