+++ app/contagens/[slug]/[compareSlug]/page.tsx
@@ -5,6 +5,7 @@ import {
   getBoxesForCountingComparision,
   getPointsDataForComparingCounting,
   getPointsDataForSingleCounting,
+  getChartData,
 } from "./configuration";
 import { VerticalStatisticsBoxes } from "../../../components/VerticalStatisticsBoxes";
 import { Map } from "../../../components/Maps/Map";
@@ -13,7 +14,9 @@ import { CountingComparisionTable, HourlyCyclistsChart } from "./useclient";
 const fetchUniqueData = async (slug: string) => {
   const id = slug.split("-")[0];
 
-  const res = await fetch(COUNTINGS_DATA_NEW + "?id=" + id, { cache: "no-cache" });
+  const res = await fetch(COUNTINGS_DATA_NEW + "?id=" + id, {
+    cache: "no-cache",
+  });
   const data = await res.json();
   return data;
 };
@@ -21,15 +24,15 @@ const fetchUniqueData = async (slug: string) => {
 const fetchData = async () => {
   const dataRes = await fetch(COUNTINGS_DATA_NEW, { cache: "no-cache" });
   const dataJson = await dataRes.json();
-  const otherData = dataJson.data;
+  const otherCounts = dataJson.data;
 
   const pageDataRes = await fetch(COUNTINGS_PAGE_DATA, { cache: "no-cache" });
   const pageCover = await pageDataRes.json();
-  return { pageCover, otherData };
+  return { pageCover, otherCounts };
 };
 
 export default async function Compare({ params }) {
-  const toCompare = [params.slug].concat(params.compareId.split("_COMPARE_"));
+  const toCompare = [params.slug].concat(params.compareSlug.split("_COMPARE_"));
   const data = await Promise.all(
     toCompare.map(async (d) => {
       const result = await fetchUniqueData(d);
@@ -37,7 +40,7 @@ export default async function Compare({ params }) {
     })
   );
 
-  const { pageCover, otherData } = await fetchData();
+  const { pageCover, otherCounts } = await fetchData();
   let pageData = {
     title: "Comparação de contagens",
     src: pageCover.cover.url,
@@ -52,43 +55,27 @@ export default async function Compare({ params }) {
 
   const crumb = {
     label: "Comparação entre contagens",
-    slug: params.compareId,
-    routes: ["/", "/contagens", params.compareId],
+    slug: params.compareSlug,
+    routes: ["/", "/contagens", params.compareSlug],
   };
 
-  // const boxes = getBoxesForCountingComparision(data);
-  // const pointsData = getPointsDataForComparingCounting(data);
+  const boxes = getBoxesForCountingComparision(data);
+  const pointsData = getPointsDataForComparingCounting(data);
+  const chartData = getChartData(data);
 
-  // const countsByHour = {};
-
-  // data.forEach((countData, index) => {
-  //   const countQuantitative = countData.data.quantitative;
-  //   Object.keys(countQuantitative).forEach((direction) => {
-  //     const directionCounts = countQuantitative[direction].count_per_hour;
-  //     Object.keys(directionCounts).forEach((hour) => {
-  //       const count = directionCounts[hour];
-  //       countsByHour[index] = countsByHour[index] || {};
-  //       countsByHour[index][hour] = (countsByHour[index][hour] || 0) + count;
-  //     });
-  //   });
-  // });
-
-  // const chartData = data.map((d, index) => ({
-  //   name: d.name,
-  //   data: Object.values(countsByHour[index]),
-  // }));
+  console.log(boxes);
 
   return (
     <main className="flex-auto">
       <NavCover {...pageData} />
       <Breadcrumb {...crumb} />
-      {/* <VerticalStatisticsBoxes
+      <VerticalStatisticsBoxes
         title={"Comparação entre as contagens"}
         boxes={boxes}
       />
       <Map pointsData={pointsData} />
       <HourlyCyclistsChart series={chartData} />
-      <CountingComparisionTable data={otherData} ids={toCompare} /> */}
+      {/*     <CountingComparisionTable data={otherCounts} ids={toCompare} /> */}
     </main>
   );
 }
